chore(client): move reverser structure out of utils
